* Add a togglebutton to switch view between graph and non-graph view * Split project in files to make reading easier
